refactor: share the filter modal scroll guard across the toolbars
